fix: duplicate tag issue for 'review-edit'
If the searched tag was one already in 'popular tags', it would show
both in 'applied tags' and 'popular tags', and thus could be added to
'applied tags' again. Now logic checks whether tag already exists in
'popular tags' and if it does, moves that over instead.

TODO: Make sure that searching for the same tag again doesn't show or
add a duplicate