fix: premature form submission & tag issue
Added jquery function to ignore 'enter' keypress for the tag search
input to not accidentally submit edit form while searching for tags.

Logic to check for duplicate in 'popular tags' was counting the event
target as a duplicate, so removing it instead of moving it. Added a
check to see whether the current parentTag matches the 'duplicate' found
first